requirements.txt
streamlit 


numpy
plotly
dash

pandas

fastapi
uvicorn
langchain
crewai
crewai-tools
langchain-openai
python-dotenv
duckduckgo-search
litellm
langchain-groq
